Remove duplication from workflow fetching function (#262)
* Remove duplication from workflow fetching function

* Update query-service.ts

* Update workflow-service.ts

* Simplify type for ValidWorkflowParameters

* Update api.d.ts
