+++ src/api.d.ts
@@ -1,0 +1,36 @@
+type WorkflowsAPIRoutePath = 'workflows' | 'workflows.archived';
+
+type WorkflowAPIRoutePath =
+  | 'workflow'
+  | 'workflow.terminate'
+  | 'events'
+  | 'query';
+
+type TaskQueueAPIRoutePath = 'task-queue';
+type ParameterlessAPIRoutePath = 'cluster' | 'settings' | 'user' | 'namespaces';
+
+type APIRoutePath =
+  | WorkflowsAPIRoutePath
+  | WorkflowAPIRoutePath
+  | ParameterlessAPIRoutePath
+  | TaskQueueAPIRoutePath
+  | WorkflowsAPIArchivalRoutePath;
+
+type APIRouteParameters = {
+  namespace: string;
+  executionId: string;
+  runId: string;
+  queue: string;
+};
+
+type WorkflowListRouteParameters = Pick<APIRouteParameters, 'namespace'>;
+
+type WorkflowRouteParameters = Pick<
+  APIRouteParameters,
+  'namespace' | 'executionId' | 'runId'
+>;
+type TaskQueueRouteParameters = Pick<APIRouteParameters, 'namespace' | 'queue'>;
+
+type ValidWorkflowEndpoints = WorkflowsAPIRoutePath;
+
+type ValidWorkflowParameters = ArchiveFilterParameters | FilterParameters;
