+++ src/lib/utilities/route-for-api.ts
@@ -1,44 +1,3 @@
-type WorkflowsAPIRoutePath =
-  | 'workflows'
-  | 'workflows.open'
-  | 'workflows.closed';
-
-type WorkflowsAPIArchivalRoutePath = 'archive';
-
-type WorkflowAPIRoutePath =
-  | 'workflow'
-  | 'workflow.terminate'
-  | 'events'
-  | 'query';
-
-type TaskQueueAPIRoutePath = 'task-queue';
-type ParameterlessAPIRoutePath = 'cluster' | 'settings' | 'user' | 'namespaces';
-type APIArchivalParameters = Pick<APIRouteParameters, 'namespace'>;
-
-type APIRoutePath =
-  | WorkflowsAPIRoutePath
-  | WorkflowAPIRoutePath
-  | ParameterlessAPIRoutePath
-  | TaskQueueAPIRoutePath
-  | WorkflowsAPIArchivalRoutePath;
-
-type APIRouteParameters = {
-  namespace: string;
-  executionId: string;
-  runId: string;
-  queue: string;
-};
-
-export type WorkflowListRouteParameters = Pick<APIRouteParameters, 'namespace'>;
-export type WorkflowRouteParameters = Pick<
-  APIRouteParameters,
-  'namespace' | 'executionId' | 'runId'
->;
-export type TaskQueueRouteParameters = Pick<
-  APIRouteParameters,
-  'namespace' | 'queue'
->;
-
 let base = (import.meta.env?.VITE_API as string) ?? process.env.VITE_API;
 if (base.endsWith('/')) base = base.slice(0, -1);
 
@@ -66,11 +25,6 @@ export function routeForApi(
   route: TaskQueueAPIRoutePath,
   parameters: TaskQueueRouteParameters,
 ): string;
-export function routeForApi(
-  route: WorkflowsAPIArchivalRoutePath,
-  parameters: APIArchivalParameters,
-): string;
-
 export function routeForApi(route: ParameterlessAPIRoutePath): string;
 export function routeForApi(
   route: APIRoutePath,
@@ -83,11 +37,9 @@ export function routeForApi(
     namespaces: '/namespaces',
     'task-queue': `/namespaces/${parameters?.namespace}/task-queues/${parameters?.queue}`,
     workflows: `/namespaces/${parameters?.namespace}/workflows`,
-    'workflows.open': `/namespaces/${parameters?.namespace}/workflows/open`,
-    'workflows.closed': `/namespaces/${parameters?.namespace}/workflows/closed`,
+    'workflows.archived': `/namespaces/${parameters?.namespace}/workflows/archived`,
     workflow: `/namespaces/${parameters?.namespace}/workflows/${parameters?.executionId}/executions/${parameters?.runId}`,
     'workflow.terminate': `/namespaces/${parameters?.namespace}/workflows/${parameters?.executionId}/executions/${parameters?.runId}/terminate`,
-    archive: `/namespaces/${parameters?.namespace}/workflows/archived`,
     events: `/namespaces/${parameters?.namespace}/workflows/${parameters?.executionId}/executions/${parameters?.runId}/events`,
     query: `/namespaces/${parameters?.namespace}/workflows/${parameters?.executionId}/executions/${parameters?.runId}/query`,
   };
